requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0
pydantic>=2.0
pyyaml>=6.0
httpx>=0.27
openai>=1.0
python-multipart>=0.0.6
